+++ src/main/java/io/github/mmalykhin/hmsproxy/routing/RoutingHandler.java
@@ -157,7 +157,7 @@ final class RoutingHandler implements InvocationHandler, NamespaceFallback {
     SpecialCaseHandler lock = new LockRoutingHandler(support, this);
     SpecialCaseHandler refreshPrivileges = new RefreshPrivilegesHandler(support);
     SpecialCaseHandler setAggrStatsFor = new SetAggrStatsForHandler(support);
-    SpecialCaseHandler alterTableReq = new AlterTableReqHandler(support, icebergTablePointerGuard);
+    SpecialCaseHandler alterTable = new AlterTableHandler(support, icebergTablePointerGuard, externalTableLocationRewriter);
     SpecialCaseHandler alterPartitionsReq = new AlterPartitionsReqHandler(support);
     SpecialCaseHandler truncateTableReq = new TruncateTableReqHandler(support);
     SpecialCaseHandler renamePartitionReq = new RenamePartitionReqHandler(support);
@@ -184,7 +184,10 @@ final class RoutingHandler implements InvocationHandler, NamespaceFallback {
         Map.entry("getTablesExt", getTablesExt),
         Map.entry("getAllMaterializedViewObjectsForRewriting", getAllMvForRewriting),
         Map.entry("set_aggr_stats_for", setAggrStatsFor),
-        Map.entry("alter_table_req", alterTableReq),
+        Map.entry("alter_table_req", alterTable),
+        Map.entry("alter_table", alterTable),
+        Map.entry("alter_table_with_environment_context", alterTable),
+        Map.entry("alter_table_with_cascade", alterTable),
         Map.entry("alter_partitions_req", alterPartitionsReq),
         Map.entry("truncate_table_req", truncateTableReq),
         Map.entry("rename_partition_req", renamePartitionReq),